+++ sdk/keyvault/azsecrets/CHANGELOG.md
@@ -19,6 +19,7 @@
   `Properties` fields, for example `SecretItem.Properties`.
 * Changed paged API content values to pointer types. For example, `ListPropertiesOfSecretsResponse.Secrets`
   changed type from `[]SecretItem` to `[]*SecretItem`.
+* `UpdateSecretProperties()` has a `Properties` parameter instead of a `Secret` parameter
 
 ### Bugs Fixed